Add Challenge 11: Calculate Total Price
In this challenge, you are given an array of product objects. Each object includes a price and quantity. Your task is to calculate the total cost of all products using the reduce() method, making sure to use an initial value. This exercise helps you practice reduce() with numeric accumulation based on object properties.